requirements.txt
aioredis==1.3.1
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
cachetools==5.2.0
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
channels==3.0.5
channels-redis==3.4.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudinary==1.29.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
daphne==3.0.2
decorator==5.1.1
Deprecated==1.2.13
dj-database-url==1.0.0
Django==4.0.7
django-celery-beat==2.3.0
django-cors-headers==3.13.0
django-filter==22.1
django-heroku==0.3.1
django-phonenumber-field==6.4.0
django-timezone-field==5.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg==1.21.3
facebook-sdk==3.1.0
Faker==14.2.0
google-api-core==2.10.0
google-api-python-client==2.60.0
google-auth==2.11.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.4
gunicorn==20.1.0
hiredis==2.0.0
httplib2==0.20.4
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsom==0.0.5
kombu==5.2.4
lark-parser==0.12.0
lxml==4.9.1
MarkupSafe==2.1.1
msgpack==1.0.4
ndg-httpsclient==0.5.1
packaging==21.3
phonenumbers==8.12.54
Pillow==9.2.0
prompt-toolkit==3.0.31
protobuf==4.21.5
psycopg2==2.9.3
pusher==3.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.4.0
pymongo==4.2.0
PyNaCl==1.5.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
python-decouple==3.6
python-json-logger==2.0.4
pytz==2022.2.1
redis==4.3.4
requests==2.28.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
service-identity==21.1.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
Twisted==22.8.0
txaio==22.2.1
typing_extensions==4.3.0
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
whitenoise==6.2.0
wrapt==1.14.1
zope.interface==5.4.0

